requirements.txt
fastapi
uvicorn
torch
pydantic
numpy
python-dotenv